Extract protectedPageServerSideProps to make creating page routes easy
